FG: STM32 SDIO ST

See merge request app-frameworks/esp_hosted!268

+++ esp_hosted_fg/host/stm32/common/util.c
@@ -17,7 +17,6 @@
 #include "util.h"
 #include "ctype.h"
 #include "string.h"
-#include "trace.h"
 
 /** Constants/Macros **/
 
@@ -82,7 +81,7 @@ int ipv4_addr_aton(const char *cp, uint32_t *ip_uint32)
 	/*
 	 * Check for trailing characters.
 	 */
-	if (*cp && (!isascii(*cp) || !isspace(*cp)))
+	if (*cp && (!isascii((uint8_t)*cp) || !isspace((uint8_t)*cp)))
 		return (0);
 	/*
 	 * Concoct the address according to
